ms_api/requirements.txt
flask
flask-Cors
flask-JWT-Extended
flask-JWT
flask-RESTful
flask-SQLAlchemy
marshmallow_sqlalchemy
python-dotenv
psycopg2-binary
google-cloud-storage
google-cloud-pubsub

